# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/salarys": {
            "get": {
                "description": "list salary entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List salary entities",
                "operationId": "list-salary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Salary"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create salary",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create salary",
                "operationId": "create-salary",
                "parameters": [
                    {
                        "description": "Salary entity",
                        "name": "salary",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.Salary"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Salary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/salarys/{id}": {
            "get": {
                "description": "get salary by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a salary entity by ID",
                "operationId": "get-salary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Salary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Salary"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get salary by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a salary entity by ID",
                "operationId": "delete-salary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Salary ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Salary": {
            "type": "object",
            "properties": {
                "accountNumber": {
                    "description": "IDEmployee\t   string",
                    "type": "string"
                },
                "assessmentID": {
                    "type": "integer"
                },
                "bonus": {
                    "type": "number"
                },
                "employeeID": {
                    "type": "integer"
                },
                "positionID": {
                    "type": "integer"
                },
                "salaryDate": {
                    "type": "string"
                },
                "salarys": {
                    "type": "number"
                }
            }
        },
...
        "ent.Assessment": {
            "type": "object",
            "properties": {
                "assessment": {
                    "description": "Assessment holds the value of the \"assessment\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AssessmentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.AssessmentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.AssessmentEdges": {
            "type": "object",
            "properties": {
                "formassessment": {
                    "description": "Formassessment holds the value of the formassessment edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Salary"
                    }
                }
            }
        },
        "ent.BeginWork": {
            "type": "object",
            "properties": {
                "BeginWork": {
                    "description": "BeginWork holds the value of the \"BeginWork\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BeginWorkQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BeginWorkEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BeginWorkEdges": {
            "type": "object",
            "properties": {
                "whenwork": {
                    "description": "Whenwork holds the value of the whenwork edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.EmployeeWorkingHours"
                    }
                }
            }
        },
...
        "ent.Day": {
            "type": "object",
            "properties": {
                "Day": {
                    "description": "Day holds the value of the \"Day\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DayQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DayEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DayEdges": {
            "type": "object",
            "properties": {
                "whatday": {
                    "description": "Whatday holds the value of the whatday edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.EmployeeWorkingHours"
                    }
                }
            }
        },
...
        "ent.Employee": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "Age holds the value of the \"age\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the EmployeeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.EmployeeEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
        "ent.EmployeeEdges": {
            "type": "object",
            "properties": {
                "employeestock": {
                    "description": "Employeestock holds the value of the employeestock edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Stock"
                    }
                },
                "formemployee": {
                    "description": "Formemployee holds the value of the formemployee edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Salary"
                    }
                },
                "whose": {
                    "description": "Whose holds the value of the whose edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.EmployeeWorkingHours"
                    }
                }
            }
        },
        "ent.EmployeeWorkingHours": {
            "type": "object",
            "properties": {
                "CodeWork": {
                    "description": "CodeWork holds the value of the \"CodeWork\" field.",
                    "type": "string"
                },
                "IDNumber": {
                    "description": "IDNumber holds the value of the \"IDNumber\" field.",
                    "type": "string"
                },
                "Wages": {
                    "description": "Wages holds the value of the \"Wages\" field.",
                    "type": "number"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the EmployeeWorkingHoursQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.EmployeeWorkingHoursEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.EmployeeWorkingHoursEdges": {
            "type": "object",
            "properties": {
                "beginwork": {
                    "description": "Beginwork holds the value of the beginwork edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BeginWork"
                },
                "day": {
                    "description": "Day holds the value of the day edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Day"
                },
                "employee": {
                    "description": "Employee holds the value of the employee edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Employee"
                },
                "getoffwork": {
                    "description": "Getoffwork holds the value of the getoffwork edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GetOffWork"
                },
                "role": {
                    "description": "Role holds the value of the role edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Role"
                }
            }
        },
        "ent.GetOffWork": {
            "type": "object",
            "properties": {
                "GetOffWork": {
                    "description": "GetOffWork holds the value of the \"GetOffWork\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GetOffWorkQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GetOffWorkEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.GetOffWorkEdges": {
            "type": "object",
            "properties": {
                "whenendwork": {
                    "description": "Whenendwork holds the value of the whenendwork edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.EmployeeWorkingHours"
                    }
                }
            }
        },
...
        "ent.Position": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PositionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PositionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "position": {
                    "description": "Position holds the value of the \"position\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PositionEdges": {
            "type": "object",
            "properties": {
                "formposition": {
                    "description": "Formposition holds the value of the formposition edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Salary"
                    }
                }
            }
        },
        "ent.Product": {
            "type": "object",
            "properties": {
                "BarcodeProduct": {
                    "description": "BarcodeProduct holds the value of the \"BarcodeProduct\" field.",
                    "type": "string"
                },
                "EXP": {
                    "description": "EXP holds the value of the \"EXP\" field.",
                    "type": "string"
                },
                "MFG": {
                    "description": "MFG holds the value of the \"MFG\" field.",
                    "type": "string"
                },
                "NameProduct": {
                    "description": "NameProduct holds the value of the \"NameProduct\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ProductQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ProductEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ProductEdges": {
            "type": "object",
            "properties": {
                "formproductonline": {
                    "description": "Formproductonline holds the value of the formproductonline edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Orderonline"
                    }
                },
                "forproduct": {
                    "description": "Forproduct holds the value of the forproduct edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Promotion"
                    }
                },
                "products": {
                    "description": "Products holds the value of the products edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Orderproduct"
                    }
                },
                "stockproduct": {
                    "description": "Stockproduct holds the value of the stockproduct edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Stock"
                    }
                }
            }
        },
...
        "ent.Role": {
            "type": "object",
            "properties": {
                "Role": {
                    "description": "Role holds the value of the \"Role\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RoleQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RoleEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Salary": {
            "type": "object",
            "properties": {
                "AccountNumber": {
                    "description": "AccountNumber holds the value of the \"AccountNumber\" field.",
                    "type": "string"
                },
                "Bonus": {
                    "description": "Bonus holds the value of the \"Bonus\" field.",
                    "type": "number"
                },
                "Salary": {
                    "description": "Salary holds the value of the \"Salary\" field.",
                    "type": "number"
                },
                "SalaryDatetime": {
                    "description": "SalaryDatetime holds the value of the \"SalaryDatetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SalaryQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SalaryEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.SalaryEdges": {
            "type": "object",
            "properties": {
                "assessment": {
                    "description": "Assessment holds the value of the assessment edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Assessment"
                },
                "employee": {
                    "description": "Employee holds the value of the employee edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Employee"
                },
                "position": {
                    "description": "Position holds the value of the position edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Position"
                }
            }
        },
        "ent.Stock": {
            "type": "object",
            "properties": {
                "Amount": {
                    "description": "Amount holds the value of the \"Amount\" field.",
                    "type": "integer"
                },
                "IDcardemployee": {
                    "description": "IDcardemployee holds the value of the \"IDcardemployee\" field.",
                    "type": "string"
                },
                "Priceproduct": {
                    "description": "Priceproduct holds the value of the \"Priceproduct\" field.",
                    "type": "number"
                },
                "Time": {
                    "description": "Time holds the value of the \"Time\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the StockQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.StockEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.StockEdges": {
            "type": "object",
            "properties": {
                "employee": {
                    "description": "Employee holds the value of the employee edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Employee"
                },
                "product": {
                    "description": "Product holds the value of the product edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Product"
                },
                "typeproduct": {
                    "description": "Typeproduct holds the value of the Typeproduct edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Typeproduct"
                },
                "zoneproduct": {
                    "description": "Zoneproduct holds the value of the zoneproduct edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Zoneproduct"
                }
            }
        },
        "ent.Typeproduct": {
            "type": "object",
            "properties": {
                "Typeproduct": {
                    "description": "Typeproduct holds the value of the \"Typeproduct\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the TypeproductQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.TypeproductEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.TypeproductEdges": {
            "type": "object",
            "properties": {
                "fromTypeproductonline": {
                    "description": "FromTypeproductonline holds the value of the fromTypeproductonline edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Orderonline"
                    }
                },
                "typeproducts": {
                    "description": "Typeproducts holds the value of the Typeproducts edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Orderproduct"
                    }
                },
                "typestock": {
                    "description": "Typestock holds the value of the typestock edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Stock"
                    }
                }
            }
        },
        "ent.Zoneproduct": {
            "type": "object",
            "properties": {
                "Zone": {
                    "description": "Zone holds the value of the \"Zone\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ZoneproductQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ZoneproductEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ZoneproductEdges": {
            "type": "object",
            "properties": {
                "zonestock": {
                    "description": "Zonestock holds the value of the zonestock edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Stock"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}